refactor: imporve runtime separation
