+++ common-types/src/lib.rs
@@ -14,18 +14,46 @@
 // You should have received a copy of the GNU General Public License
 // along with Unique Network. If not, see <http://www.gnu.org/licenses/>.
 
+#![cfg_attr(not(feature = "std"), no_std)]
+
 use sp_runtime::{
-	traits::{Verify, IdentifyAccount, BlakeTwo256},
-	generic, MultiSignature,
+    generic,
+	traits::{Verify, IdentifyAccount}, MultiSignature,
 };
 
-pub use sp_runtime::OpaqueExtrinsic as UncheckedExtrinsic;
+/// Opaque types. These are used by the CLI to instantiate machinery that don't need to know
+/// the specifics of the runtime. They can then be made to be agnostic over specific formats
+/// of data like extrinsics, allowing for them to continue syncing the network through upgrades
+/// to even the core data structures.
+pub mod opaque {
+    pub use sp_runtime::{
+        generic,
+        traits::BlakeTwo256,
+        OpaqueExtrinsic as UncheckedExtrinsic
+    };
 
-/// Opaque block header type.
-pub type Header = generic::Header<BlockNumber, BlakeTwo256>;
+	pub use super::{BlockNumber, Signature, AccountId, Balance, Index, Hash, AuraId};
 
-/// Opaque block type.
-pub type Block = generic::Block<Header, UncheckedExtrinsic>;
+    /// Opaque block header type.
+    pub type Header = generic::Header<BlockNumber, BlakeTwo256>;
+
+    /// Opaque block type.
+    pub type Block = generic::Block<Header, UncheckedExtrinsic>;
+
+    pub trait RuntimeInstance {
+        type CrossAccountId: pallet_evm::account::CrossAccountId<sp_runtime::AccountId32>
+            + Send
+            + Sync
+            + 'static;
+
+        type TransactionConverter: fp_rpc::ConvertTransaction<UncheckedExtrinsic>
+            + Send
+            + Sync
+            + 'static;
+
+        fn get_transaction_converter() -> Self::TransactionConverter;
+    }
+}
 
 pub type SessionHandlers = ();
 
@@ -56,17 +84,3 @@ pub type Hash = sp_core::H256;
 pub type DigestItem = generic::DigestItem;
 
 pub use sp_consensus_aura::sr25519::AuthorityId as AuraId;
-
-pub trait RuntimeInstance {
-	type CrossAccountId: pallet_evm::account::CrossAccountId<sp_runtime::AccountId32>
-		+ Send
-		+ Sync
-		+ 'static;
-
-	type TransactionConverter: fp_rpc::ConvertTransaction<UncheckedExtrinsic>
-		+ Send
-		+ Sync
-		+ 'static;
-
-	fn get_transaction_converter() -> Self::TransactionConverter;
-}
