+++ runtime/common/ethereum/mod.rs
@@ -14,13 +14,12 @@
 // You should have received a copy of the GNU General Public License
 // along with Unique Network. If not, see <http://www.gnu.org/licenses/>.
 
-#![cfg_attr(not(feature = "std"), no_std)]
-
-pub mod constants;
-pub mod construct_runtime;
-pub mod dispatch;
-pub mod eth_sponsoring;
-pub mod runtime_apis;
 pub mod sponsoring;
-pub mod types;
-pub mod weights;
+pub mod transaction_converter;
+pub mod self_contained_call;
+
+pub use {
+    sponsoring::*,
+    transaction_converter::*,
+    self_contained_call::*,
+};
