+++ runtime/opal/Cargo.toml
@@ -113,13 +113,15 @@ std = [
     'xcm/std',
     'xcm-builder/std',
     'xcm-executor/std',
-    'unique-runtime-common/std',
+    'common-types/std',
     'rmrk-rpc/std',
+    'evm-coder/std',
+    'up-sponsorship/std',
 
     "orml-vesting/std",
 ]
 limit-testing = ['pallet-unique/limit-testing', 'up-data-structs/limit-testing']
-opal-runtime = ['scheduler', 'rmrk']
+opal-runtime = ['rmrk', 'scheduler']
 
 scheduler = []
 rmrk = []
@@ -400,7 +402,7 @@ default-features = false
 
 [dependencies]
 log = { version = "0.4.16", default-features = false }
-unique-runtime-common = { path = "../common", default-features = false, features = ['refungible'] }
+common-types = { path = "../../common-types", default-features = false }
 scale-info = { version = "2.0.1", default-features = false, features = [
     "derive",
 ] }
@@ -430,6 +432,8 @@ pallet-ethereum = { default-features = false, git = "https://github.com/uniquene
 pallet-base-fee = { default-features = false, git = "https://github.com/uniquenetwork/frontier", branch = "unique-polkadot-v0.9.24" }
 fp-rpc = { default-features = false, git = "https://github.com/uniquenetwork/frontier", branch = "unique-polkadot-v0.9.24" }
 fp-self-contained = { default-features = false, git = "https://github.com/uniquenetwork/frontier", branch = "unique-polkadot-v0.9.24" }
+evm-coder = { default-features = false, path = '../../crates/evm-coder' }
+up-sponsorship = { default-features = false, git = "https://github.com/uniquenetwork/pallet-sponsoring", branch = 'polkadot-v0.9.24' }
 
 ################################################################################
 # Build Dependencies
